IMPALA-10340: Cannot set up KDC from scratch
IMPALA-9361 added experimental support for setting up a KDC for the mini
cluster.  However, the experimental-kerberos-setup.sh script fails with
the error:

  KRB5_CONFIG: unbound variable

The problem is that impala-config.sh unsets KDC-related variables if the
cluster is not kerberized or does not exist.  However, kerberized
cluster cannot be created without installing KDC first.

This change updates impala-config.sh to allow setting the KDC-related
variables if IMPALA_KERBERIZE is set to 'true' and the cluster does not
exist yet.

Testing:
 * successfully set up KDC via experimental-kerberos-setup.sh after
   building Impala locally
 * verified that impala-config.sh sets/unsets KDC-related variables
   depending on cluster existence and config
 * started Kerberized and non-Kerberized mini clusters (after
   regenerating configs)

Change-Id: Icaed2ad2fb1e1b60951a5a4138a9386588bc6972